chore: rename map component
